+++ tests/Unit/Headers/GlobalHeadersTest.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 it('can include global headers for every request')
     ->fake()
-    ->tap(fn () => $this->soap()->headers($this->soap()->header('Auth', 'test.com', ['foo' => 'bar'])))
+    ->defer(fn () => $this->soap()->headers($this->soap()->header('Auth', 'test.com', ['foo' => 'bar'])))
     ->soap()->to(EXAMPLE_SOAP_ENDPOINT)->call('Add', ['intB' => 25])
     ->test()->assertSent(fn ($request) => $request->getHeaders() == [
         $this->soap()->header('Auth', 'test.com', ['foo' => 'bar']),
@@ -12,8 +12,8 @@ it('can include global headers for every request')
 
 it('can include scoped headers based on the endpoint')
     ->fake()
-    ->tap(fn () => $this->soap()->headers($this->soap()->header('Brand', 'test.coms', ['hello' => 'world']))->for('https://foo.bar'))
-    ->tap(fn () => $this->soap()->headers($this->soap()->header('Auth', 'test.com', ['foo' => 'bar']))->for(EXAMPLE_SOAP_ENDPOINT))
+    ->defer(fn () => $this->soap()->headers($this->soap()->header('Brand', 'test.coms', ['hello' => 'world']))->for('https://foo.bar'))
+    ->defer(fn () => $this->soap()->headers($this->soap()->header('Auth', 'test.com', ['foo' => 'bar']))->for(EXAMPLE_SOAP_ENDPOINT))
     ->soap()->to(EXAMPLE_SOAP_ENDPOINT)->call('Add', ['intB' => 25])
     ->test()->assertSent(fn ($request) => $request->getHeaders() == [
         $this->soap()->header('Auth', 'test.com', ['foo' => 'bar']),
@@ -21,8 +21,8 @@ it('can include scoped headers based on the endpoint')
 
 it('can scope headers based on the endpoint and method')
     ->fake()
-    ->tap(fn () => $this->soap()->headers($this->soap()->header('Brand', 'test.coms', ['hello' => 'world']))->for(EXAMPLE_SOAP_ENDPOINT, 'Add'))
-    ->tap(fn () => $this->soap()->headers($this->soap()->header('Auth', 'test.com', ['foo' => 'bar']))->for(EXAMPLE_SOAP_ENDPOINT, 'Subtract'))
+    ->defer(fn () => $this->soap()->headers($this->soap()->header('Brand', 'test.coms', ['hello' => 'world']))->for(EXAMPLE_SOAP_ENDPOINT, 'Add'))
+    ->defer(fn () => $this->soap()->headers($this->soap()->header('Auth', 'test.com', ['foo' => 'bar']))->for(EXAMPLE_SOAP_ENDPOINT, 'Subtract'))
     ->soap()->to(EXAMPLE_SOAP_ENDPOINT)->call('Add', ['intB' => 25])
     ->test()->assertSent(fn ($request) => $request->getHeaders() == [
         $this->soap()->header('Brand', 'test.coms', ['hello' => 'world']),
@@ -30,8 +30,8 @@ it('can scope headers based on the endpoint and method')
 
 it('merges the global headers with the local headers')
     ->fake()
-    ->tap(fn () => $this->soap()->headers($this->soap()->header('Brand', 'test.coms', ['hello' => 'world'])))
-    ->tap(fn () => $this->soap()
+    ->defer(fn () => $this->soap()->headers($this->soap()->header('Brand', 'test.coms', ['hello' => 'world'])))
+    ->defer(fn () => $this->soap()
         ->to(EXAMPLE_SOAP_ENDPOINT)
         ->withHeaders($this->soap()->header('Auth', 'test.com', ['foo' => 'bar']))
         ->call('Add', ['intB' => 25])