+++ tests/Unit/Headers/HeadersTest.php
@@ -6,7 +6,7 @@ use RicorocksDigitalAgency\Soap\Request\SoapPhpRequest;
 
 it('can set headers')
     ->fake()
-    ->tap(fn () => $this->soap()->to(EXAMPLE_SOAP_ENDPOINT)
+    ->defer(fn () => $this->soap()->to(EXAMPLE_SOAP_ENDPOINT)
         ->withHeaders($this->soap()->header('Auth', 'test.com')->data(['foo' => 'bar']))
         ->call('Add', ['intA' => 10, 'intB' => 25])
     )
@@ -16,7 +16,7 @@ it('can set headers')
 
 it('can define multiple headers in the same method')
     ->fake()
-    ->tap(fn () => $this->soap()->to(EXAMPLE_SOAP_ENDPOINT)
+    ->defer(fn () => $this->soap()->to(EXAMPLE_SOAP_ENDPOINT)
         ->withHeaders(
             $this->soap()->header('Auth', 'test.com')->data(['foo' => 'bar']),
             $this->soap()->header('Brand', 'test.com')->data(['hello' => 'world'])
@@ -30,7 +30,7 @@ it('can define multiple headers in the same method')
 
 it('can define multiple headers with an array in the same method')
     ->fake()
-    ->tap(fn () => $this->soap()->to(EXAMPLE_SOAP_ENDPOINT)
+    ->defer(fn () => $this->soap()->to(EXAMPLE_SOAP_ENDPOINT)
         ->withHeaders(...[
             $this->soap()->header('Auth', 'test.com')->data(['foo' => 'bar']),
             $this->soap()->header('Brand', 'test.com')->data(['hello' => 'world']),
@@ -44,7 +44,7 @@ it('can define multiple headers with an array in the same method')
 
 it('can define multiple headers using a collection in the same method')
     ->fake()
-    ->tap(fn () => $this->soap()->to(EXAMPLE_SOAP_ENDPOINT)
+    ->defer(fn () => $this->soap()->to(EXAMPLE_SOAP_ENDPOINT)
         ->withHeaders(...collect([
             $this->soap()->header('Auth', 'test.com')->data(['foo' => 'bar']),
             $this->soap()->header('Brand', 'test.com')->data(['hello' => 'world']),
@@ -58,7 +58,7 @@ it('can define multiple headers using a collection in the same method')
 
 it('can define multiple headers in multiple methods')
     ->fake()
-    ->tap(fn () => $this->soap()->to(EXAMPLE_SOAP_ENDPOINT)
+    ->defer(fn () => $this->soap()->to(EXAMPLE_SOAP_ENDPOINT)
         ->withHeaders($this->soap()->header('Auth', 'test.com')->data(['foo' => 'bar']))
         ->withHeaders($this->soap()->header('Brand', 'test.com')->data(['hello' => 'world']))
         ->call('Add', ['intA' => 10, 'intB' => 25])
@@ -70,7 +70,7 @@ it('can define multiple headers in multiple methods')
 
 it('can create a header without any parameters and be composed fluently')
     ->fake()
-    ->tap(fn () => $this->soap()->to(EXAMPLE_SOAP_ENDPOINT)->withHeaders(
+    ->defer(fn () => $this->soap()->to(EXAMPLE_SOAP_ENDPOINT)->withHeaders(
         $this->soap()->header()
             ->name('Auth')
             ->namespace('test.com')
@@ -86,7 +86,7 @@ it('can create a header without any parameters and be composed fluently')
 
 it('can set up a header using a SoapVar')
     ->fake()
-    ->tap(fn () => $this->soap()->to(EXAMPLE_SOAP_ENDPOINT)
+    ->defer(fn () => $this->soap()->to(EXAMPLE_SOAP_ENDPOINT)
         ->withHeaders($this->soap()->header('Auth', 'test.com', new SoapVar(['foo' => 'bar'], null)))
         ->call('Add', ['intA' => 10, 'intB' => 25])
     )
@@ -96,7 +96,7 @@ it('can set up a header using a SoapVar')
 
 it('does not require the data parameter')
     ->fake()
-    ->tap(fn () => $this->soap()->to(EXAMPLE_SOAP_ENDPOINT)
+    ->defer(fn () => $this->soap()->to(EXAMPLE_SOAP_ENDPOINT)
         ->withHeaders($this->soap()->header('Auth', 'test.com')->data(null))
         ->withHeaders($this->soap()->header('Brand', 'test.com', null))
         ->call('Add', ['intA' => 10, 'intB' => 25])